UPD: Modularized Structure of Scoreboard
